requirements.txt
transformers
datasets
torch
scikit-learn
pandas
flask
flask-socketio
nltk
spacy
azure-cognitiveservices-speech
flask-cors
